requirements.txt
pymupdf==1.26.3
qdrant-client==1.15.1
fastembed==0.7.1
streamlit==1.48.0
python-dotenv==1.1.1
asyncio==4.0.0
pandas==2.3.1
numpy==2.3.2
sentence-transformers== 5.0.0